+++ apps/opplaringspenger-soknad/mock/data/index.ts
@@ -1,0 +1,11 @@
+import { arbeidsgiverMock } from './arbeidsgiverMock';
+import { barnMock } from './barnMock';
+import { institusjonerMock } from './opplæringsinstitusjonMock';
+import { søkerMock } from './søkerMock';
+
+export const mockData = {
+    barn: barnMock,
+    søker: søkerMock,
+    arbeidsgiver: arbeidsgiverMock,
+    institusjoner: institusjonerMock,
+};