+++ packages/sif-common-amplitude/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/sif-common-amplitude
 
+## 2.28.12
+
+### Patch Changes
+
+- Pakkeoppdateringer
+
+## 2.28.11
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/major + zod4
+
 ## 2.28.10
 
 ### Patch Changes